requirements.txt
anthropic
typer
rich
python-dotenv
pydantic>=2,<3

